etl_pipeline/requirements.txt
# Core
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0
# IO
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pyarrow>=16.0.0
fastparquet>=2024.5.0
# Reporting
Jinja2>=3.1.3
# Optional (only if you plan to generate PDF from HTML; tooling like wkhtmltopdf is also required)
# pdfkit>=1.0.0
